pyproject.toml
[project]
name = "neteasemusicconverter"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "pillow==12.0.0",
    "pycryptodome==3.23.0",
]
